Ai_app/edge-fastapi-inference/edge_inference/requirements.txt
fastapi
uvicorn
pillow
numpy
python-multipart
requests
ultralytics

